Dropped 'Java-like' support
Dropped official support for Java, Ant and Android.
The modes themselves didn't do a lot and were very old. I'm planning to revive
these modes in the future, maybe using JDEE as base or simply creating a lot
of wrappers and relying on ant/maven tasks for compiling. No concrete plans for
this though.