Major update on all custom files
Every custom file was reviewed to remove the use of `el-get'; instead the
function `install-if-needed' (to be defined in the main file) is used.
Also, `defshell' was replaced with `eshell' and the .gitignore was updated
to ignore eshell internal directory.

The best of this update is that Windows users will have a much easier time,
since they're relying only on package.el and not a bunch of other scattered
tools.